% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indiator_means.R
\name{calculate_ind_longterm_mean}
\alias{calculate_ind_longterm_mean}
\title{To calculate the mean and sd of the total timeseries for the indicator}
\usage{
calculate_ind_longterm_mean(data, area, start_year, end_year, ind)
}
\arguments{
\item{data}{is the ecosystem indicator data}

\item{area}{is the management area}

\item{ind}{is the column of the indicator}
}
\description{
To calculate the mean and sd of the total timeseries for the indicator
}
