% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indiator_means.R
\name{calculate_ind_period_mean}
\alias{calculate_ind_period_mean}
\title{Calculates the mean and sd of a specific time period}
\usage{
calculate_ind_period_mean(data, area, start_year, end_year, ind)
}
\arguments{
\item{ind}{}
}
\description{
Calculates the mean and sd of a specific time period
}
