% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_diversity.R
\name{plot_catch_diversity}
\alias{plot_catch_diversity}
\title{Plot Shannon diversity of catches}
\usage{
plot_catch_diversity(
  data,
  x_column,
  y_column,
  color_var,
  y_label,
  legend_labels
)
}
\arguments{
\item{x_column}{variable that is typically year}

\item{y_column}{variable that is the calculated shannon index}

\item{color_var}{character or factor variable}

\item{y_label}{character string in ""}

\item{legend_labels}{list of legend labels in ""}
}
\value{
a plot with smoothed lines and an averaged loess line
}
\description{
Plot Shannon diversity of catches
}
