% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_indicators.R
\name{plot_multiple_indicators}
\alias{plot_multiple_indicators}
\title{Plot multiple indicator time series that fit into themes}
\usage{
plot_multiple_indicators(
  data,
  x_column,
  y_column,
  color_var,
  y_label,
  legend_labels
)
}
\arguments{
\item{data}{is a data frame with YEAR, value, name}

\item{x_column}{is the YEAR column}

\item{y_column}{is the value column}

\item{color_var}{is the name column}

\item{y_label}{is a character string in ""}

\item{legend_labels}{is a group of character strings in ""}
}
\value{
a plot
}
\description{
Plot multiple indicator time series that fit into themes
}
