% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nao.R
\name{plot_nao}
\alias{plot_nao}
\title{plotting function for NAO}
\usage{
plot_nao(data, x_column, y_column, mean_value, sd_value, y_label)
}
\arguments{
\item{data}{nao data derived from .dat files}

\item{x_column}{year column}

\item{y_column}{nao values}

\item{mean_value}{mean derived from nao list in dd}

\item{sd_value}{sd derived from nao list in dd}

\item{y_label}{label for nao}
}
\value{
line plot with NAO mean and sd
}
\description{
plotting function for NAO
}
