% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature_anomaly.R
\name{plot_temperature_anomaly}
\alias{plot_temperature_anomaly}
\title{Plotting temperature anomalies with rigime shifts}
\usage{
plot_temperature_anomaly(
  data,
  x_column,
  y_column,
  mean_value,
  sd_value,
  ff,
  bb,
  y_label
)
}
\arguments{
\item{data}{temperature data derived from .dat files}

\item{x_column}{year column}

\item{y_column}{temperature anomaly}

\item{mean_value}{mean set at 0}

\item{sd_value}{sd value set at 0.5}

\item{ff}{regime shift analysis using forward routine}

\item{bb}{regime shift analysis using backwards routime}

\item{y_label}{}
}
\value{
plot with mean, sd, regime shift lines
}
\description{
Plotting temperature anomalies with rigime shifts
}
