% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plotting function for time series indicators}
\usage{
plot_indicator(data, x_column, y_column)
}
\arguments{
\item{data}{is a data frame with YEAR and Indicator value}

\item{x_column}{YEAR column}

\item{y_column}{Indicator column}
}
\description{
Plotting function for time series indicators
}
