% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_ref_period.R
\name{plot_indicator_ref_period}
\alias{plot_indicator_ref_period}
\title{plotting function for indicators with reference points for 30 years consistent with climate data}
\usage{
plot_indicator_ref_period(data, x_column, y_column, y_label)
}
\arguments{
\item{data}{indicator data}

\item{x_column}{year column}

\item{y_column}{indicator value column}

\item{y_label}{indicator name as character string "}
}
\value{
a line plot with points, mean+/-sd
}
\description{
plotting function for indicators with reference points for 30 years consistent with climate data
}
